Fix: Add missing _tp parameter to specialized REF output builders
Added _tp=self.value_tp parameter to make_instance() methods in:
- PythonTSREFOutputBuilder (REF[TS[...]])
- PythonTSDREFOutputBuilder (REF[TSD[...]])
- PythonTSSREFOutputBuilder (REF[TSS[...]])
- PythonTSWREFOutputBuilder (REF[TSW[...]])

This ensures consistency with the generic PythonREFOutputBuilder and
the specialized TSL/TSB builders, all of which pass the _tp parameter.
The _tp field is defined in the base PythonTimeSeriesReferenceOutput
class and may be used for type checking or debugging.

All tests pass with both Python and C++ runtimes.